+++ src/MyJetWallet.MatchingEngine.EventReader/MatchingEngineCashOutEventReader.cs
@@ -28,7 +28,7 @@ namespace MyJetWallet.MatchingEngine.EventReader
         protected override string ExchangeName { get; }
         protected override string QueueName { get; }
         protected override bool IsQueueAutoDelete { get; }
-        protected override string[] RoutingKeys { get; } = { MessageType.CashOut.ToString() };
+        protected override string[] RoutingKeys { get; } = { ((int)MessageType.CashOut).ToString() };
 
         protected override async Task ProcessBatch(IList<CustomQueueItem<CashOutEvent>> batch)
         {
