+++ src/MyJetWallet.MatchingEngine.EventReader/MatchingEngineReservedBalanceUpdateEventReader.cs
@@ -28,7 +28,7 @@ namespace MyJetWallet.MatchingEngine.EventReader
         protected override string ExchangeName { get; }
         protected override string QueueName { get; }
         protected override bool IsQueueAutoDelete { get; }
-        protected override string[] RoutingKeys { get; } = { MessageType.ReservedBalanceUpdate.ToString() };
+        protected override string[] RoutingKeys { get; } = { ((int)MessageType.ReservedBalanceUpdate).ToString() };
 
         protected override async Task ProcessBatch(IList<CustomQueueItem<ReservedBalanceUpdateEvent>> batch)
         {
